moved all deprecated functions and tests
